requirements.txt
requests>=2.31.0
Pillow>=10.0.0
streamlit>=1.28.0
